define workflows with manual input validation
