+++ convex/workflows/engine.ts
@@ -3,26 +3,25 @@ import { v } from 'convex/values'
 import { internal } from '../_generated/api'
 import { internalAction, internalMutation } from '../functions'
 import { getErrorMessage } from '../shared/utils'
-import { textToImageWorkflow } from './textToImage'
-import { WorkflowType } from './types'
+import { textToImagePipeline } from './pipelines/textToImage'
 
-const workflows: { [K in WorkflowType['type']]: WorkflowType } = {
-  textToImage: textToImageWorkflow,
+const pipelines = {
+  textToImage: textToImagePipeline,
 }
 
-export const startWorkflow = internalMutation({
+export const startJob = internalMutation({
   args: {
-    workflowType: v.string(),
+    pipeline: v.string(),
     input: v.any(),
   },
-  handler: async (ctx, args): Promise<string> => {
-    const workflow = workflows[args.workflowType as WorkflowType['type']]
-    if (!workflow) {
-      throw new Error(`Unknown workflow type: ${args.workflowType}`)
+  handler: async (ctx, args) => {
+    const pipeline = pipelines[args.pipeline as keyof typeof pipelines]
+    if (!pipeline) {
+      throw new Error(`Unknown pipeline: ${args.pipeline}`)
     }
 
-    const jobId = await ctx.table('jobs2').insert({
-      workflowType: args.workflowType,
+    const jobId = await ctx.table('jobs3').insert({
+      pipeline: args.pipeline,
       status: 'pending',
       currentStep: 0,
       input: args.input,
@@ -30,69 +29,88 @@ export const startWorkflow = internalMutation({
       updatedAt: Date.now(),
     })
 
-    await ctx.scheduler.runAfter(0, internal.workflows.engine.executeNextStep, { jobId })
-    console.log('job created', args.workflowType, jobId)
+    await ctx.scheduler.runAfter(0, internal.workflows.engine.executeStep, { jobId })
+    console.log('job created', args.pipeline, jobId)
     return jobId
   },
 })
 
-export const executeNextStep = internalAction({
+export const executeStep = internalAction({
   args: {
-    jobId: v.id('jobs2'),
+    jobId: v.id('jobs3'),
   },
-
-  handler: async (ctx, { jobId }): Promise<void> => {
+  handler: async (ctx, { jobId }) => {
     const job = await ctx.runQuery(internal.workflows.jobs.get, { jobId })
-    if (!job) throw new Error('job not found')
+    if (!job) {
+      throw new Error('job not found')
+    }
 
-    const workflow = workflows[job.workflowType as WorkflowType['type']]
-    const currentStep = workflow.steps[job.currentStep]
-    if (!currentStep) {
-      await ctx.runMutation(internal.workflows.jobs.updateStatus, { jobId, status: 'completed' })
+    const pipeline = pipelines[job.pipeline as keyof typeof pipelines]
+    const step = pipeline.steps[job.currentStep]
+    // * no step - assume job finished
+    if (!step) {
+      await ctx.runMutation(internal.workflows.jobs.updateStatus, {
+        jobId,
+        status: 'completed',
+      })
       return
     }
 
-    console.log('execute next step', job.currentStep, currentStep.name, jobId)
+    const stepStartTime = Date.now()
     try {
+      // * run step
       const previousResults = job.stepResults.map((r) => r.result)
-      const result = await currentStep.action(ctx, job.input, previousResults)
-      console.log('engine step result', result)
+      const result = await step.action(ctx, job.input, previousResults)
 
+      // * add successful result
       await ctx.runMutation(internal.workflows.jobs.addStepResult, {
         jobId,
         stepResult: {
-          stepName: currentStep.name,
+          stepName: step.name,
           status: 'completed',
           result,
-          startTime: Date.now(),
+          startTime: stepStartTime,
           endTime: Date.now(),
           retryCount: 0,
         },
       })
 
-      // * Schedule next step / return
-      await ctx.scheduler.runAfter(0, internal.workflows.engine.executeNextStep, { jobId })
-    } catch (error) {
-      const retryCount = (job.stepResults[job.currentStep]?.retryCount ?? 0) + 1
-      if (retryCount > currentStep.retryLimit) {
-        // * max retries reached
-        await ctx.runMutation(internal.workflows.jobs.updateStatus, { jobId, status: 'failed' })
+      // * schedule next step, return
+      await ctx.scheduler.runAfter(0, internal.workflows.engine.executeStep, { jobId })
+    } catch (err) {
+      const retryCount = (job.stepResults.at(-1)?.retryCount ?? 0) + 1
+
+      if (retryCount > step.retryLimit) {
+        // * max retries reached, fail
+        await ctx.runMutation(internal.workflows.jobs.updateStatus, {
+          jobId,
+          status: 'failed',
+        })
       } else {
+        // * add failed step error, retry
         await ctx.runMutation(internal.workflows.jobs.addStepResult, {
           jobId,
           stepResult: {
-            stepName: currentStep.name,
+            stepName: step.name,
             status: 'failed',
-            error: getErrorMessage(error),
-            result: undefined,
-            startTime: Date.now(),
+            error: {
+              code: 'job_error',
+              message: getErrorMessage(err),
+              fatal: false,
+            },
+            result: null,
+            startTime: stepStartTime,
             endTime: Date.now(),
             retryCount,
           },
         })
 
         // * Retry after a delay
-        await ctx.scheduler.runAfter(5000, internal.workflows.engine.executeNextStep, { jobId })
+        await ctx.scheduler.runAfter(
+          2000 * (retryCount + 1),
+          internal.workflows.engine.executeStep,
+          { jobId },
+        )
       }
     }
   },
