+++ convex/workflows/jobs.ts
@@ -1,29 +1,42 @@
 import { v } from 'convex/values'
 
 import { internalMutation, internalQuery } from '../functions'
-import { job2Fields } from '../schema'
+import { job3Fields } from '../schema'
 
 export const get = internalQuery({
   args: {
-    jobId: v.id('jobs2'),
+    jobId: v.id('jobs3'),
   },
-  handler: async (ctx, { jobId }) => {
-    return await ctx.table('jobs2').get(jobId)
+  handler: async (ctx, args) => {
+    return await ctx.table('jobs3').get(args.jobId)
+  },
+})
+
+export const updateStatus = internalMutation({
+  args: {
+    jobId: v.id('jobs3'),
+    status: job3Fields.status,
+  },
+  handler: async (ctx, { jobId, status }) => {
+    const job = await ctx.table('jobs3').getX(jobId)
+    return await job.patch({
+      status,
+    })
   },
 })
 
 export const addStepResult = internalMutation({
   args: {
-    jobId: v.id('jobs2'),
-    stepResult: job2Fields.stepResults.element,
+    jobId: v.id('jobs3'),
+    stepResult: job3Fields.stepResults.element,
   },
 
   handler: async (ctx, { jobId, stepResult }) => {
-    const job = await ctx.table('jobs2').getX(jobId)
+    const job = await ctx.table('jobs3').getX(jobId)
     console.log('add step result', stepResult)
     try {
       return await job.patch({
-        status: 'in_progress',
+        status: 'active',
         stepResults: [...job.stepResults, stepResult],
         currentStep: stepResult.status === 'completed' ? job.currentStep + 1 : job.currentStep,
         updatedAt: Date.now(),
@@ -32,18 +45,5 @@ export const addStepResult = internalMutation({
       console.log(stepResult)
       throw err
     }
-  },
-})
-
-export const updateStatus = internalMutation({
-  args: {
-    jobId: v.id('jobs2'),
-    status: v.string(),
-  },
-  handler: async (ctx, { jobId, status }) => {
-    const job = await ctx.table('jobs2').getX(jobId)
-    return await job.patch({
-      status,
-    })
   },
 })
